Extract packet processing from mixnode-common
