+++ common/nymsphinx/framing/src/lib.rs
@@ -3,3 +3,4 @@
 
 pub mod codec;
 pub mod packet;
+pub mod processing;
