+++ gateway/src/node/mixnet_handling/receiver/packet_processing.rs
@@ -3,18 +3,24 @@
 
 use nym_crypto::asymmetric::encryption;
 use nym_mixnode_common::packet_processor::error::MixProcessingError;
-pub use nym_mixnode_common::packet_processor::processor::MixProcessingResult;
-use nym_mixnode_common::packet_processor::processor::{ProcessedFinalHop, SphinxPacketProcessor};
+use nym_mixnode_common::packet_processor::processor::SphinxPacketProcessor;
 use nym_sphinx::framing::packet::FramedNymPacket;
+use nym_sphinx::framing::processing::{
+    process_framed_packet, MixProcessingResult, PacketProcessingError, ProcessedFinalHop,
+};
+use nym_sphinx::PrivateKey;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum GatewayProcessingError {
     #[error("failed to process received mix packet - {0}")]
-    PacketProcessingError(#[from] MixProcessingError),
+    PacketProcessing(#[from] MixProcessingError),
 
     #[error("received a forward hop mix packet")]
-    ForwardHopReceivedError,
+    ForwardHopReceived,
+
+    #[error("failed to process received sphinx packet: {0}")]
+    NymPacketProcessing(#[from] PacketProcessingError),
 }
 
 // PacketProcessor contains all data required to correctly unwrap and store sphinx packets
@@ -24,21 +30,23 @@ pub struct PacketProcessor {
 }
 
 impl PacketProcessor {
+    pub fn sphinx_key(&self) -> &PrivateKey {
+        self.inner_processor.sphinx_key()
+    }
+
     pub(crate) fn new(encryption_key: &encryption::PrivateKey) -> Self {
         PacketProcessor {
             inner_processor: SphinxPacketProcessor::new(encryption_key.into()),
         }
     }
+}
 
-    pub(crate) fn process_received(
-        &self,
-        received: FramedNymPacket,
-    ) -> Result<ProcessedFinalHop, GatewayProcessingError> {
-        match self.inner_processor.process_received(received)? {
-            MixProcessingResult::ForwardHop(..) => {
-                Err(GatewayProcessingError::ForwardHopReceivedError)
-            }
-            MixProcessingResult::FinalHop(processed_final) => Ok(processed_final),
-        }
+pub(crate) fn process_packet(
+    received: FramedNymPacket,
+    sphinx_key: &nym_sphinx::PrivateKey,
+) -> Result<ProcessedFinalHop, GatewayProcessingError> {
+    match process_framed_packet(received, sphinx_key)? {
+        MixProcessingResult::ForwardHop(..) => Err(GatewayProcessingError::ForwardHopReceived),
+        MixProcessingResult::FinalHop(processed_final) => Ok(processed_final),
     }
 }
